Made it so only the user who created an order or item is able to delete or update it

+++ partout/Backend/src/main/java/com/websites/partout/resource/ItemResource.java
@@ -2,8 +2,12 @@ package com.websites.partout.resource;
 
 import com.websites.partout.dao.ItemRepo;
 import com.websites.partout.model.Item;
+import com.websites.partout.utils.JwtUtil;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -18,6 +22,9 @@ public class ItemResource {
     @Autowired
     ItemRepo itemRepo;
 
+    @Autowired
+    JwtUtil jwtUtil;
+
     @GetMapping
     public List<Item> getAll() {
         return itemRepo.findAll();
@@ -25,9 +32,9 @@ public class ItemResource {
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public void persist(@RequestBody final Item order) {
+    public void persist(@RequestBody final Item item) {
         try {
-            itemRepo.save(order);
+            itemRepo.save(item);
         }
         catch (Exception ex) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Bad or incomplete item information");
@@ -47,38 +54,67 @@ public class ItemResource {
 
     @DeleteMapping(path = "/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteItemById(@PathVariable("id") final int id) {
+    public void deleteItemById(@PathVariable("id") final int id, @RequestHeader("Authorization") HttpHeaders headers) {
         try {
-            itemRepo.deleteById(id);
+            final List<String> token = headers.get("Authorization");
+            if (token.get(0).startsWith("Bearer ")) {
+                String jwt = token.get(0).substring(6);
+                Item item = itemRepo.findById(id).orElse(null);
+                int callerId = jwtUtil.extractCallerId(jwt);
+                if (item != null && item.fk_user_id == callerId) {
+                    itemRepo.deleteById(id);
+                } else {
+                    throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You can only delete your own items");
+                }
+            }
+            else {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "You have not logged in");
+            }
+        }
+        catch (ResponseStatusException ex) {
+            throw new ResponseStatusException(ex.getStatus(), ex.getReason());
         }
         catch (Exception ex) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Invalid item id");
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
         }
     }
 
     @PutMapping(path = "/{id}")
-    public Optional<Item> updateUserById(@PathVariable("id") final int id, @RequestBody final Item item) {
+    public Optional<Item> updateItemById(@PathVariable("id") final int id, @RequestBody final Item item, @RequestHeader("Authorization") HttpHeaders headers) {
         Optional<Item> tempItem = itemRepo.findById(id);
         if (tempItem.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid item id");
         }
         else {
-            itemRepo.findById(id).map(u -> {
-                int indexOfUser = u.getId_Item();
-                if (indexOfUser > 0) {
-                    try {
-                    itemRepo.deleteById(indexOfUser);
-                    item.setId_Item(id);
-                    return itemRepo.save(item);
-                    }
-                    catch (Exception ex) {
-                        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid item id");
-                    }
+            final List<String> token = headers.get("Authorization");
+            if (token.get(0).startsWith("Bearer ")) {
+                String jwt = token.get(0).substring(6);
+                Item itemTemp = itemRepo.findById(id).orElse(null);
+                int callerId = jwtUtil.extractCallerId(jwt);
+                if (itemTemp != null && itemTemp.fk_user_id == callerId) {
+                    itemRepo.findById(id).map(u -> {
+                        int indexOfItem = u.getId_Item();
+                        if (indexOfItem > 0) {
+                            try {
+                                itemRepo.deleteById(indexOfItem);
+                                item.setId_Item(id);
+                                return itemRepo.save(item);
+                            }
+                            catch (Exception ex) {
+                                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid item id");
+                            }
+                        }
+                        else {
+                            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid item id");
+                        }
+                    });
+                } else {
+                    throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You can only update your own items");
                 }
-                else {
-                    throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid item id");
-                }
-            });
+            }
+            else {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "You have not logged in");
+            }
             throw new ResponseStatusException(HttpStatus.FOUND);
         }
     }
